src/main/java/io/github/vlad324/n226/Solution.java
package io.github.vlad324.n226;

import java.util.Objects;

/**
 * {@link "https://leetcode.com/problems/invert-binary-tree/"}
 */
class Solution {

    // TODO: iterative approach

    public TreeNode invertTree(TreeNode root) {
        if (root == null) {
            return null;
        }

        invertTree(root.left);
        invertTree(root.right);

        return swap(root);
    }

    private TreeNode swap(TreeNode root) {
        final var temp = root.left;
        root.left = root.right;
        root.right = temp;

        return root;
    }


    static class TreeNode {
        int val;
        TreeNode left;
        TreeNode right;

        TreeNode(int val) {
            this.val = val;
        }

        TreeNode(int val, TreeNode left, TreeNode right) {
            this.val = val;
            this.left = left;
            this.right = right;
        }

        @Override
        public boolean equals(Object o) {
            if (this == o) return true;
            if (o == null || getClass() != o.getClass()) return false;
            final TreeNode treeNode = (TreeNode) o;
            return val == treeNode.val &&
                Objects.equals(left, treeNode.left) &&
                Objects.equals(right, treeNode.right);
        }

        @Override
        public int hashCode() {
            return Objects.hash(val, left, right);
        }
    }

}
